chore: spliting discover page
