+++ packages/web/components/discover/discover-filter-section.component.tsx
@@ -1,0 +1,18 @@
+import React, { ReactNode } from 'react';
+import { DiscoverFilterSectionStyles } from './discover-filter-section.styles';
+
+interface DiscoverFilterSectionComponentProps {
+  title: string;
+  children: ReactNode;
+}
+export function DiscoverFilterSectionComponent(
+  props: DiscoverFilterSectionComponentProps
+) {
+  const { title, children } = props;
+  return (
+    <DiscoverFilterSectionStyles>
+      <h6>{title}</h6>
+      {children}
+    </DiscoverFilterSectionStyles>
+  );
+}
