+++ packages/web/components/navbar/navbar.component.tsx
@@ -11,6 +11,7 @@ const links = [
   ['TV Shows', '/library/tvshows'],
   ['Search', '/search'],
   ['Discover', '/discover'],
+  ['Suggestions', '/suggestions'],
   ['Settings', '/settings'],
 ];
